Ingreso y retiro del codigo.

+++ app/Http/Controllers/OtpController.php
@@ -5,27 +5,39 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\OtpService;
 use App\Otp;
+use App\Store;
 
 class OtpController extends Controller
 {
-    public function create()
+    public function create($idstore, $idclient)
     {
-        $otp = $this->store();
+        $otp = $this->store($idstore, $idclient);
         return view('otps.buy', ['otp' => $otp]);
     }
 
-    public function store()
+    public function store($store_id, $client_id)
     {
         $otpService = new OtpService();
-        $otp = $otpService->generateOtp(1);
+        $otp = $otpService->generateOtp($client_id, $store_id);
         return $otp;
     }
 
-    public function destroy(Request $request)
+    public function destroy(Request $request, $idstore)
     {
-        $id = $request->codigo;
-        $otp = Otp::find($id);
-        $otp->delete();
-        return back();
+        $store = Store::where('user_id', $idstore)->first();
+        $pass = $request->codigo;
+        $otp = Otp::where('otp_pass', $pass)->first();
+        $message = "";
+        if ($otp != null && $store != null) {
+            if ($otp->store_id == $store->id) {
+                $otp->delete();
+                $message = "Codigo ingresado correctamente!";
+            } else {
+                $message = "Ups! El codigo o el comercio es incorrecto";
+            }
+        } else {
+            $message = "Ups! El codigo o el comercio es incorrecto";
+        }
+        return back()->with('info', $message);
     }
 }
